added red swipe on delete posts with delete icon # Conflicts: #	app/src/main/java/com/example/seniorproject/data/Firebase/FirebaseData.kt #	app/src/main/java/com/example/seniorproject/data/repositories/PostRepository.kt

+++ app/src/main/java/com/example/seniorproject/MainForum/Posts/CommunityPosts.kt
@@ -1,8 +1,14 @@
 package com.example.seniorproject.MainForum.Posts
 
+import android.content.DialogInterface
+import android.graphics.Canvas
+import android.graphics.Color
+import android.graphics.drawable.ColorDrawable
+import android.graphics.drawable.Drawable
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
+import androidx.appcompat.app.AlertDialog
 import androidx.core.content.ContextCompat
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
@@ -23,6 +29,10 @@ import javax.inject.Inject
 class CommunityPosts : AppCompatActivity() {
 
     private lateinit var adapter: CustomAdapter
+    var swipeBackground: ColorDrawable = ColorDrawable(Color.parseColor("#FF0000"))
+    lateinit var deleteIcon: Drawable
+    var builder = AlertDialog.Builder(this)
+
 
     @Inject
     lateinit var factory: ViewModelProvider.Factory
@@ -62,19 +72,66 @@ class CommunityPosts : AppCompatActivity() {
 
         }
 
+        deleteIcon = ContextCompat.getDrawable(this, R.drawable.ic_delete_24px)!!
+
+
         val itemTouchHelperCallback =
             object : ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT) {
                 override fun onSwiped(viewHolders: RecyclerView.ViewHolder, position: Int) {
                     val postkey: String? =
                         adapter.removeItem(viewHolders as CustomViewHolders, position)
 
+                    builder.setTitle("Delete Post")
+                    builder.setMessage("Are you sure you want to delete this post?")
+                    builder. setPositiveButton("YES", {dialogInterface: DialogInterface?, i: Int ->
+                        myViewModel.deletePost(postkey!!, className)
+                    })
+                    builder. setNegativeButton("NO", {dialogInterface: DialogInterface?, i: Int ->
+                        builder.show()
+                    })
                     //.getStringExtra("Classkey")
                     //val postkey = intent.getStringExtra("author")
-                    myViewModel.deletePost(postkey!!, className)
+                    //myViewModel.deletePost(postkey!!, className)
                     //myViewModel.deletePost()
 
                     //  obse.onChanged(pos)
                 }
+
+                override fun onChildDraw(
+               c: Canvas,
+               recyclerView: RecyclerView,
+               viewHolder: RecyclerView.ViewHolder,
+               dX: Float,
+               dY: Float,
+               actionState: Int,
+               isCurrentlyActive: Boolean
+           ) {
+               val itemView = viewHolder.itemView
+               val iconMargin = (itemView.height - deleteIcon.intrinsicHeight)/2
+
+               if(dX > 0) {
+                  // swipeBackground.setBounds(itemView.left, itemView.top, dX.toInt(), itemView.bottom)
+                  // deleteIcon.setBounds(itemView.left+ iconMargin, itemView.top + iconMargin, itemView.left +
+                  // iconMargin + deleteIcon.intrinsicWidth, itemView.bottom - iconMargin)
+               } else{
+                   swipeBackground.setBounds(itemView.right + dX.toInt(), itemView.top, itemView.right, itemView.bottom)
+                   deleteIcon.setBounds(itemView.right - iconMargin - deleteIcon.intrinsicWidth, itemView.top + iconMargin,
+                       itemView.right - iconMargin, itemView.bottom - iconMargin)
+               }
+
+               swipeBackground.draw(c)
+               deleteIcon.draw(c)
+
+               super.onChildDraw(
+                   c,
+                   recyclerView,
+                   viewHolder,
+                   dX,
+                   dY,
+                   actionState,
+                   isCurrentlyActive
+               )
+           }
 
                 override fun onMove(
                     recyclerView: RecyclerView,
@@ -87,6 +144,8 @@ class CommunityPosts : AppCompatActivity() {
             }
 
         val itemTouchHelper = ItemTouchHelper(itemTouchHelperCallback)
+
+
         itemTouchHelper.attachToRecyclerView(classes_post_RV)
 
     }
